+++ WinCopies.IO/ShellObjectInfo.cs
@@ -356,17 +356,17 @@ namespace WinCopies.IO
 
                 if (FileType == FileType.Folder || FileType == FileType.Drive || FileType == FileType.SpecialFolder)
 
-                    LoadItems(new FolderLoader(workerReportsProgress, workerSupportsCancellation, FileTypes.All));
+                    LoadItems(new FolderLoader(workerReportsProgress, workerSupportsCancellation, GetAllEnumFlags<FileTypes>()));
 
                 else if (FileType == FileType.Archive)
 
-                    LoadItems(new ArchiveLoader(workerReportsProgress, workerSupportsCancellation, FileTypes.All));
+                    LoadItems(new ArchiveLoader(workerReportsProgress, workerSupportsCancellation, GetAllEnumFlags<FileTypes>()));
 
             }
 
             else
 
-                LoadItems(new FolderLoader(workerReportsProgress, workerSupportsCancellation, FileTypes.All));
+                LoadItems(new FolderLoader(workerReportsProgress, workerSupportsCancellation, GetAllEnumFlags<FileTypes>()));
 
             //else
 
@@ -404,17 +404,17 @@ namespace WinCopies.IO
 
                 if (FileType == FileType.Folder || FileType == FileType.Drive || FileType == FileType.SpecialFolder)
 
-                    LoadItemsAsync(new FolderLoader(workerReportsProgress, workerSupportsCancellation, FileTypes.All));
+                    LoadItemsAsync(new FolderLoader(workerReportsProgress, workerSupportsCancellation, GetAllEnumFlags<FileTypes>()));
 
                 else if (FileType == FileType.Archive)
 
-                    LoadItemsAsync(new ArchiveLoader(workerReportsProgress, workerSupportsCancellation, FileTypes.All));
+                    LoadItemsAsync(new ArchiveLoader(workerReportsProgress, workerSupportsCancellation, GetAllEnumFlags<FileTypes>()));
 
             }
 
             else
 
-                LoadItemsAsync(new FolderLoader(workerReportsProgress, workerSupportsCancellation, FileTypes.All));
+                LoadItemsAsync(new FolderLoader(workerReportsProgress, workerSupportsCancellation, GetAllEnumFlags<FileTypes>()));
 
             //else